refactor: raycasting
