+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: caide-so <caide-so@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/08 19:37:37 by caide-so          #+#    #+#             */
-/*   Updated: 2025/07/16 22:57:23 by caide-so         ###   ########.fr       */
+/*   Updated: 2025/07/17 21:16:22 by caide-so         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -182,5 +182,8 @@ int		render_frame(t_game *game);
 
 // raycast
 void	raycast_and_draw(t_game *game);
+void	init_ray(t_player *player, t_ray *ray, int x);
+void	perform_dda(t_ray *ray, t_game *game);
+void	draw_column(t_ray *ray, t_game *game, int x);
 
 #endif
