+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: caide-so <caide-so@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/08 19:21:33 by caide-so          #+#    #+#             */
-/*   Updated: 2025/07/16 20:57:26 by caide-so         ###   ########.fr       */
+/*   Updated: 2025/07/17 21:51:02 by caide-so         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
